Merge Spring Security y reservas

See merge request ctd/proyecto-integrador-0822/1021-ft/grupo-11!32

+++ BackEnd/ProyectoJava/PI-BackEnd/src/main/java/com/ctd/proyectointegrador/persistance/repository/CaracteristicaRepository.java
@@ -3,5 +3,5 @@ package com.ctd.proyectointegrador.persistance.repository;
 import com.ctd.proyectointegrador.persistance.model.Caracteristica;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CaracteristicaRepository extends JpaRepository<Caracteristica, Integer> {
+public interface CaracteristicaRepository extends JpaRepository<Caracteristica, Long> {
 }
